backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
celery>=5.4
redis>=5.0
django-celery-results>=2.5

numpy>=2.1
scipy>=1.13
librosa>=0.10.2
soundfile>=0.12
noisereduce>=3.0
pydub>=0.25

python-dotenv>=1.0